# caseframesvc/swagger.yml
swagger: '2.0'
info:
  title: Case Frame Service API
  description: Provide access to Address Frame and Response Management information
  version: 8.11.0
schemes:
  - http
basePath: /frameservice
produces:
  - application/json
...
paths:
...
  /samples:
    get:
      summary: Get all samples
      description: Return details of all samples.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
      tags:
        - Response
      responses:
        '200':
          description: An array of samples
          schema:
            type: array
            items:
              $ref: '#/definitions/Sample'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /samples/{sampleid}:
    put:
      summary: Create cases for the specified sample.
      description: |
        Update the supplied sample by creating the cases for the provided geography type and code.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: sampleid
          in: path
          description: Unique sample Id.
          required: true
          type: integer
        - name: geographyDetails
          in: body
          description: Geography for which to create cases.
          required: true
          schema:
            $ref: '#/definitions/GeographyDetails'
      tags:
        - Response
      responses:
        '200':
          description: Cases created.
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /surveys:
    get:
      summary: Get all surveys.
      description: Return details of all surveys.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
      tags:
        - Response
      responses:
        '200':
          description: An array of surveys.
          schema:
            type: array
            items:
              $ref: '#/definitions/Survey'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /surveys/{surveyid}:
    get:
      summary: Get information for the specified survey.
      description: Return information for the supplied survey code.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: surveyid
          in: path
          description: Unique survey Id.
          required: true
          type: integer
      tags:
        - Response
      responses:
        '200':
          description: A single survey.
          schema:
              $ref: '#/definitions/Survey'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Sample:
    type: object
    properties:
      sampleId:
        type: integer
        description: Unique Id of sample.
      sampleName:
        type: string
        description: Name of sample.
      sampleDescription:
        type: string
        description: Description of sample.
      addressCriteria:
        type: string
        description: Address selection criteria.
      caseTypeId:
        type: integer
        description: Unique Id of case type.
      surveyId:
        type: integer
        description: Survey instrument.
  Survey:
    type: object
    properties:
      surveyId:
        type: integer
        description: Unique Id of survey.
      surveyName:
        type: string
        description: Name of survey.
      description:
        type: string
        description: Description of survey.
...
  GeographyDetails:
    type: object
    properties:
      geographyType:
        description: Type of geography area for which to create cases e.g. REGION, LA, MSOA, LSOA, OA.
        type: string
      geographyCode:
        description: Geography area code value for which to create cases.
        type: string
  Error:
    type: object
    properties:
      code:
        type: string
      timestamp:
        type: string
      message:
        type: string